Rename the join planner module to hash_join
The planner in `core/src/plan/join.rs` detects hash join candidates and
produces `JoinExecutorPlan::Hash` plans, preserving the nested-loop plan
when no candidate is found. Rename the module so its name matches that
responsibility.

No logic is changed; the file content is identical to before.
