+++ core/src/plan.rs
@@ -2,8 +2,8 @@ mod aggregate;
 mod context;
 mod error;
 mod expr;
+mod hash_join;
 mod index;
-mod join;
 mod planner;
 mod primary_key;
 mod schema;
@@ -13,6 +13,6 @@ mod validate;
 
 pub use {
     self::validate::validate, aggregate::plan as plan_aggregate, error::*, expr::plan_scalar_expr,
-    index::plan as plan_index, join::plan as plan_join, primary_key::plan as plan_primary_key,
+    hash_join::plan as plan_join, index::plan as plan_index, primary_key::plan as plan_primary_key,
     schema::fetch_schema_map, schemaless::plan as plan_schemaless, statement::*,
 };
